+++ changes.md
@@ -1,3 +1,13 @@
+### 0.6.2
+* added changes to controller.ssh.entrypoint. Fixes for exec command
+
+### 0.6.1
+* added changes to fix startup warnings
+
+### 0.6.0
+* updated kubectl to 1.32.0
+* changed kubectl package source
+
 ### 0.5.5
 * OS updates
 * Updated node modules. Updated Alpine to node:23.5-alpine